Update README.md
# <Weather Dashboard>

#Description

A web dashboard built to show the current weather in a city of choice and the upcoming five day forecast for that city

Enter in your city of choice and select search to view the current weather, temperature, humidity, wind speed, and UV Index for that city at the current time. Upon the search, the dashobard will also display the weather, temperature and hummidity forecast for the next five(5) days.

UV Index will be accompanied by a color indicating the severity of the index. Favorable is shown in gree, Moderate is shown in yellow and severe is shown in red.

Previous searches are saved in localStorage and the five (5) most recent searches are available as buttons in order for easy access and re-searching of weather and forecasts


Built With
HTML
CSS
Javascript
Open Weather APIs - Current Weather Data, UV Index and 5 day/ 3 Hour Forecast
Bootstrap API
Moment.js API


![The weather app includes a search option, a list of cities, and a five-day forecast and current weather conditions for Atlanta.](./Users/desaradoko/Downloads/UNH-POR-VIRT-FSF-PT-12-2021-U-C-main 3/01-Class-Content/06-Server-Side-APIs/02-Homework/Assets/06-server-side-apis-homework-demo.png)